Upload schema endpoint; auto-update schema on deploy; fix ingestion number serialization use integer for integers

+++ dataspray-stream-ingest/src/test/java/io/dataspray/stream/ingest/IngestBase.java
@@ -222,7 +222,7 @@ public abstract class IngestBase extends AbstractLambdaTest {
 
         Object messageTsObject = objectJson.get(ETL_MESSAGE_TS);
         log.info("Object ts {}", messageTsObject);
-        Instant messageTsInstant = Instant.ofEpochMilli(((Double) messageTsObject).longValue());
+        Instant messageTsInstant = Instant.ofEpochMilli(((long) messageTsObject));
         assertTrue(messageTsInstant.isAfter(Instant.now().minusSeconds(100)));
         assertTrue(messageTsInstant.isBefore(Instant.now().plusSeconds(100)));
